# openapi.yaml
openapi: 3.1.0
info:
  title: Titipin API
  version: 0.2.0
  description: |
    Titipin menyediakan jasa titip belanja lokal antara pemesan dan jastiper.
    Protected operations require the declared OAuth2 scope and a separate object relationship check.
    All monetary amounts are integers in the minor unit
    of the stated currency. Clients MUST ignore unknown
    response fields.
servers:
  - url: https://api.titipin.example/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - oauth2: [requests:read]
paths:
...
  /requests/{requestId}/offers:
    get:
      tags: [Offers]
      operationId: listRequestOffers
      security:
        - oauth2: [requests:read]
      summary: List offers for a request
      description: Returns a paginated collection of offers belonging to the request. The status filter only narrows offers.
      x-required-roles: [requester, jastiper, admin]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OfferStatusQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Paginated offer collection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferPage'
              example:
                items: []
                page:
                  limit: 10
                  nextCursor: null
                  hasMore: false
        '400': { $ref: '#/components/responses/InvalidRequest400' }
        '401': { $ref: '#/components/responses/AuthenticationRequired401' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/ResourceNotFound404' }
        '500': { $ref: '#/components/responses/InternalError500' }
        '503': { $ref: '#/components/responses/ServiceUnavailable503' }
    post:
      tags: [Offers]
      operationId: createOffer
      security:
        - oauth2: [requests:fulfil]
      summary: Submit an offer
      description: |
        Creates an offer for an open request. Only a verified jastiper may submit an offer.
        Stock is an observation at stockCheckedAt, not a managed store inventory resource.
      x-required-roles: [jastiper]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOffer'
            example:
              itemPrice:
                amount: 120000
                currency: IDR
              serviceFee:
                amount: 15000
                currency: IDR
              deliveryFee:
                amount: 10000
                currency: IDR
              estimatedArrivalAt: '2026-09-04T18:00:00+07:00'
              stockCheckedAt: '2026-09-03T08:10:00+07:00'
              expiresAt: '2026-09-03T12:00:00+07:00'
              note: Item is available at the specified store.
      responses:
        '201':
          description: Offer created.
          headers:
            Location:
              description: URI of the created offer.
              schema:
                type: string
                example: /v1/offers/off_9Zx28LmQ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
              example:
                offerId: off_9Zx28LmQ
                requestId: req_51Fa93cD
                jastiperId: acc_jas_2Rt71BnP
                status: active
                itemPrice:
                  amount: 120000
                  currency: IDR
                serviceFee:
                  amount: 15000
                  currency: IDR
                deliveryFee:
                  amount: 10000
                  currency: IDR
                totalAmount:
                  amount: 145000
                  currency: IDR
                estimatedArrivalAt: '2026-09-04T18:00:00+07:00'
                stockCheckedAt: '2026-09-03T08:10:00+07:00'
                expiresAt: '2026-09-03T12:00:00+07:00'
                note: Item is available at the specified store.
                createdAt: '2026-09-03T08:10:00+07:00'
        '400': { $ref: '#/components/responses/PostBadRequest400' }
        '401': { $ref: '#/components/responses/AuthenticationRequired401' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/ResourceNotFound404' }
        '409': { $ref: '#/components/responses/InvalidStateOrIdempotency409' }
        '422': { $ref: '#/components/responses/ValidationFailed422' }
        '412': { $ref: '#/components/responses/PreconditionFailed412' }
        '500': { $ref: '#/components/responses/InternalError500' }
        '503': { $ref: '#/components/responses/ServiceUnavailable503' }
  /offers/{offerId}:
    get:
      tags: [Offers]
      operationId: getOffer
      security:
        - oauth2: [requests:read]
      summary: Get an offer
      description: Returns one offer representation.
      x-required-roles: [requester, jastiper, admin]
      parameters:
        - $ref: '#/components/parameters/OfferId'
      responses:
        '200':
          description: Offer representation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
              example:
                offerId: off_9Zx28LmQ
                requestId: req_51Fa93cD
                jastiperId: acc_jas_2Rt71BnP
                status: active
                itemPrice:
                  amount: 120000
                  currency: IDR
                serviceFee:
                  amount: 15000
                  currency: IDR
                deliveryFee:
                  amount: 10000
                  currency: IDR
                totalAmount:
                  amount: 145000
                  currency: IDR
                estimatedArrivalAt: '2026-09-04T18:00:00+07:00'
                stockCheckedAt: '2026-09-03T08:10:00+07:00'
                expiresAt: '2026-09-03T12:00:00+07:00'
                note: Item is available at the specified store.
                createdAt: '2026-09-03T08:10:00+07:00'
        '401': { $ref: '#/components/responses/AuthenticationRequired401' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/ResourceNotFound404' }
        '500': { $ref: '#/components/responses/InternalError500' }
        '503': { $ref: '#/components/responses/ServiceUnavailable503' }
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        Access tokens are issued by the Titipin authorisation server. Public
        clients use Authorization Code + PKCE S256 with no client secret.
        URLs below describe the Step 3 local realm; publish the reachable HTTPS
        issuer URLs before using this document with a deployed client.
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8081/realms/titipin/protocol/openid-connect/auth
          tokenUrl: http://localhost:8081/realms/titipin/protocol/openid-connect/token
          refreshUrl: http://localhost:8081/realms/titipin/protocol/openid-connect/token
          scopes:
            requests:read: Read requests, offers, assignments, deliveries, and receipt confirmations visible to the caller
            requests:write: Create and manage your own requests - place a request, select an offer, pay, and confirm receipt
            requests:fulfil: Jastiper operations to fulfil an assigned request - submit an offer and record a purchase and delivery
            deliveries:write: Submit delivery location updates, as a jastiper or a tracking device
            payments:read: Read payment records
            accounts:read: Read limited account representations
            issues:read: Read transaction issues assigned to the administrator or MCP reader
            issues:write: Report own transaction issues or resolve assigned issues as an administrator
        clientCredentials:
          tokenUrl: http://localhost:8081/realms/titipin/protocol/openid-connect/token
          scopes:
            requests:read: Read transaction context explicitly assigned to the MCP service
            issues:read: Read issues explicitly assigned to the MCP service
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Canonical hyphenated UUID version 4. Generate it when the user confirms the action and reuse it unchanged with the same method, URI, and body for every retry of that intent.
        The service retains the key and stored result for 24 hours from first receipt. Reuse with a different request returns 409 idempotency-key-reuse. An identical request still in progress returns 409 with Retry-After.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
        example: 0f7c1b9e-3d21-4a6f-9c05-8e2b7d41a9f0
    RequestId:
      name: requestId
      in: path
      required: true
      description: Opaque server-generated request identifier.
      example: req_51Fa93cD
      schema:
        type: string
        minLength: 8
        maxLength: 64
        example: req_51Fa93cD
    OfferId:
      name: offerId
      in: path
      required: true
      description: Opaque server-generated offer identifier.
      example: off_9Zx28LmQ
      schema:
        type: string
        minLength: 8
        maxLength: 64
        example: off_9Zx28LmQ
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return. The service applies a default and an upper bound.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 10
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page.
      schema:
        type: string
        maxLength: 512
        example: eyJvZmZzZXQiOjEwfQ
...
    OfferStatusQuery:
      name: status
      in: query
      required: false
      description: Filters offers by one known status.
      schema:
        type: string
        enum: [active, selected, withdrawn, expired]
        example: active
...
  responses:
    InvalidRequest400:
      description: The request could not be parsed or its query, header, or structure is invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            invalidRequest:
              summary: Invalid request
              value:
                type: 'https://api.titipin.example/problems/invalid-request'
                title: Invalid request
                status: 400
                detail: The request body could not be parsed as valid JSON.
                instance: /v1/requests
                invalidParameters:
                  - name: body
                    location: body
                    reason: malformed JSON
    PostBadRequest400:
      description: The POST request is invalid or its required Idempotency-Key is missing or malformed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            invalidRequest:
              summary: Invalid request
              value:
                type: 'https://api.titipin.example/problems/invalid-request'
                title: Invalid request
                status: 400
                detail: The request body could not be parsed as valid JSON.
                instance: /v1/requests
                invalidParameters:
                  - name: body
                    location: body
                    reason: malformed JSON
            invalidIdempotencyKey:
              summary: Invalid idempotency key
              value:
                type: 'https://api.titipin.example/problems/invalid-idempotency-key'
                title: Invalid idempotency key
                status: 400
                detail: Idempotency-Key must be a canonical UUID version 4.
                instance: /v1/requests
                headerName: Idempotency-Key
    AuthenticationRequired401:
      description: The access token is missing, unverifiable, expired, or has the wrong issuer or audience.
      headers:
        WWW-Authenticate:
          description: Authentication challenge.
          schema:
            type: string
            example: 'Bearer error="invalid_token"'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: 'https://api.titipin.example/problems/authentication-required'
            title: Authentication required
            status: 401
            detail: A valid access token is required.
            instance: /v1/requests
    Forbidden403:
      description: The token is valid but lacks the operation scope; no object is loaded.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          example: 'Bearer error="insufficient_scope", scope="requests:write"'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: 'https://api.titipin.example/problems/forbidden'
            title: Action is not permitted
            status: 403
            detail: The authenticated account does not have permission to perform this action.
            instance: /v1/requests
            requiredPermission: requests:write
    PreconditionFailed412:
      description: The client wrote using an ETag that is no longer current.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: 'https://api.titipin.example/problems/precondition-failed'
            title: The resource changed
            status: 412
            detail: Data ini sudah berubah karena diproses lebih dulu oleh pengguna lain.
            instance: /v1/requests/req_51Fa93cD
    ResourceNotFound404:
      description: |
        The object is absent or inaccessible to this caller. Both conditions
        use the same status, type, title, detail and constant problem instance.
        Resource identifiers and ownership hints are never included.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: 'https://api.titipin.example/problems/resource-not-found'
            title: Resource not found
            status: 404
            detail: The requested resource was not found.
            instance: urn:titipin:problem:resource-not-found
    ValidationFailed422:
      description: The JSON is valid, but one or more values violate a schema or domain constraint.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: 'https://api.titipin.example/problems/validation-failed'
            title: Request validation failed
            status: 422
            detail: One or more request fields are invalid.
            instance: /v1/requests
            errors:
              - field: quantity
                code: minimum
                message: quantity must be greater than or equal to 1
...
    InvalidStateOrIdempotency409:
      description: The operation is not allowed in the current state, or the idempotency key conflicts with an earlier request.
      headers:
        Retry-After:
          description: Present only for idempotency-request-in-progress.
          schema:
            type: integer
            minimum: 1
            example: 2
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            invalidStateTransition:
              summary: Invalid state transition
              value:
                type: 'https://api.titipin.example/problems/invalid-state-transition'
                title: State transition is not allowed
                status: 409
                detail: The resource is not in a state that permits this operation.
                instance: /v1/requests/req_51Fa93cD/offers
                resourceType: Request
                resourceId: req_51Fa93cD
                currentStatus: assigned
                allowedStatuses: [open]
            idempotencyKeyReuse:
              summary: Idempotency key reused for a different request
              value:
                type: 'https://api.titipin.example/problems/idempotency-key-reuse'
                title: Idempotency key reused for a different request
                status: 409
                detail: The supplied Idempotency-Key is already bound to another request.
                instance: /v1/requests/req_51Fa93cD/offers
                originalRequestUri: /v1/requests/req_51Fa93cD/offers
            idempotencyRequestInProgress:
              summary: Original request still in progress
              value:
                type: 'https://api.titipin.example/problems/idempotency-request-in-progress'
                title: The original request is still being processed
                status: 409
                detail: Retry the same request after the number of seconds stated in Retry-After.
                instance: /v1/requests/req_51Fa93cD/offers
                retryAfterSeconds: 2
...
    InternalError500:
      description: Unexpected internal service failure without internal implementation details.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: 'https://api.titipin.example/problems/internal-error'
            title: An internal error occurred
            status: 500
            detail: The service could not complete the request.
            instance: /v1/requests
    ServiceUnavailable503:
      description: The service or an important dependency is temporarily unavailable. Retry with exponential backoff and jitter.
      headers:
        Retry-After:
          description: Suggested minimum delay before retrying, when available.
          schema:
            type: integer
            minimum: 1
            example: 10
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: 'https://api.titipin.example/problems/service-unavailable'
            title: Service is temporarily unavailable
            status: 503
            detail: The service is temporarily unable to process the request.
            instance: /v1/requests
            retryAfterSeconds: 10
  schemas:
...
    Money:
      type: object
      description: Monetary value represented as an integer minor-unit amount and currency code.
      required: [amount, currency]
      properties:
        amount:
          type: integer
          minimum: 0
          description: Amount in the minor unit of the stated currency. For IDR this is the integer rupiah amount.
          example: 150000
        currency:
          type: string
          enum: [IDR]
          description: ISO 4217 currency code for the amount.
          example: IDR
      example:
        amount: 150000
        currency: IDR
    PageInfo:
      type: object
      required: [limit, nextCursor, hasMore]
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of items requested for the page.
          example: 10
        nextCursor:
          type: [string, 'null']
          description: Opaque cursor for the next page, or null when there is no next page.
          example: eyJvZmZzZXQiOjEwfQ
        hasMore:
          type: boolean
          description: Whether another page is available.
          example: true
      example:
        limit: 10
        nextCursor: eyJvZmZzZXQiOjEwfQ
        hasMore: true
...
    OfferStatus:
      type: string
      enum: [active, selected, withdrawn, expired]
      description: Clients MUST treat an unrecognised value as non-terminal and MUST NOT crash.
      example: active
...
    Offer:
      type: object
      additionalProperties: true
      required: [offerId, requestId, jastiperId, status, itemPrice, serviceFee, deliveryFee, totalAmount, estimatedArrivalAt, stockCheckedAt, expiresAt, note, createdAt]
      properties:
        offerId:
          type: string
          description: Opaque server-generated offer identifier.
          example: off_9Zx28LmQ
        requestId:
          type: string
          description: Request to which the offer belongs.
          example: req_51Fa93cD
        jastiperId:
          type: string
          description: Account identifier of the verified jastiper who submitted the offer.
          example: acc_jas_2Rt71BnP
        status:
          $ref: '#/components/schemas/OfferStatus'
          description: Current offer lifecycle status.
          example: active
        itemPrice:
          $ref: '#/components/schemas/Money'
          description: Expected purchase price of the requested item.
          example:
            amount: 120000
            currency: IDR
        serviceFee:
          $ref: '#/components/schemas/Money'
          description: Simulated jastiper service fee.
          example:
            amount: 15000
            currency: IDR
        deliveryFee:
          $ref: '#/components/schemas/Money'
          description: Simulated delivery fee.
          example:
            amount: 10000
            currency: IDR
        totalAmount:
          $ref: '#/components/schemas/Money'
          description: Sum of item price, service fee, and delivery fee.
          example:
            amount: 145000
            currency: IDR
        estimatedArrivalAt:
          type: string
          format: date-time
          description: Estimated delivery arrival time with timezone offset.
          example: '2026-09-04T18:00:00+07:00'
        stockCheckedAt:
          type: string
          format: date-time
          description: Time at which the jastiper observed stock availability.
          example: '2026-09-03T08:10:00+07:00'
        expiresAt:
          type: string
          format: date-time
          description: Time after which the offer is no longer active.
          example: '2026-09-03T12:00:00+07:00'
        note:
          type: [string, 'null']
          maxLength: 500
          description: Optional human-readable note about the offer. No stock or purchase evidence is included.
          example: Item is available at the specified store.
        createdAt:
          type: string
          format: date-time
          description: RFC 3339 offer creation timestamp with timezone offset.
          example: '2026-09-03T08:10:00+07:00'
      example:
        offerId: off_9Zx28LmQ
        requestId: req_51Fa93cD
        jastiperId: acc_jas_2Rt71BnP
        status: active
        itemPrice:
          amount: 120000
          currency: IDR
        serviceFee:
          amount: 15000
          currency: IDR
        deliveryFee:
          amount: 10000
          currency: IDR
        totalAmount:
          amount: 145000
          currency: IDR
        estimatedArrivalAt: '2026-09-04T18:00:00+07:00'
        stockCheckedAt: '2026-09-03T08:10:00+07:00'
        expiresAt: '2026-09-03T12:00:00+07:00'
        note: Item is available at the specified store.
        createdAt: '2026-09-03T08:10:00+07:00'
    CreateOffer:
      type: object
      additionalProperties: true
      required: [itemPrice, serviceFee, deliveryFee, estimatedArrivalAt, stockCheckedAt, expiresAt]
      properties:
        itemPrice:
          $ref: '#/components/schemas/Money'
          description: Expected purchase price of the requested item.
          example:
            amount: 120000
            currency: IDR
        serviceFee:
          $ref: '#/components/schemas/Money'
          description: Simulated jastiper service fee.
          example:
            amount: 15000
            currency: IDR
        deliveryFee:
          $ref: '#/components/schemas/Money'
          description: Simulated delivery fee.
          example:
            amount: 10000
            currency: IDR
        estimatedArrivalAt:
          type: string
          format: date-time
          description: Estimated delivery arrival time with timezone offset.
          example: '2026-09-04T18:00:00+07:00'
        stockCheckedAt:
          type: string
          format: date-time
          description: Time at which the jastiper observed stock availability.
          example: '2026-09-03T08:10:00+07:00'
        expiresAt:
          type: string
          format: date-time
          description: Time after which the offer is no longer active.
          example: '2026-09-03T12:00:00+07:00'
        note:
          type: [string, 'null']
          maxLength: 500
          description: Optional human-readable note about the offer.
          example: Item is available at the specified store.
      example:
        itemPrice:
          amount: 120000
          currency: IDR
        serviceFee:
          amount: 15000
          currency: IDR
        deliveryFee:
          amount: 10000
          currency: IDR
        estimatedArrivalAt: '2026-09-04T18:00:00+07:00'
        stockCheckedAt: '2026-09-03T08:10:00+07:00'
        expiresAt: '2026-09-03T12:00:00+07:00'
        note: Item is available at the specified store.
...
    OfferPage:
      type: object
      additionalProperties: true
      required: [items, page]
      properties:
        items:
          type: array
          description: Offers in the page.
          items:
            $ref: '#/components/schemas/Offer'
          example: []
        page:
          $ref: '#/components/schemas/PageInfo'
          description: Pagination metadata.
          example:
            limit: 10
            nextCursor: null
            hasMore: false
      example:
        items: []
        page:
          limit: 10
          nextCursor: null
          hasMore: false
...